Refactor - Create class MaterialYouNavigationBarCustom | SetupScreen navigationBar
