+++ lib/screens/setupScreen/setup_screen.dart
@@ -1,15 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_banking_pay_responsive/constant_text_styles.dart';
-import 'package:flutter_banking_pay_responsive/constants.dart';
+import 'package:flutter_banking_pay_responsive/components/material_you_navigation_bar_custom.dart';
 import 'package:flutter_banking_pay_responsive/core/android_quick_actions_shotcuts.dart';
 import 'package:flutter_banking_pay_responsive/data_providers.dart';
-import 'package:flutter_banking_pay_responsive/generated/l10n.dart';
 import 'package:flutter_banking_pay_responsive/main.dart';
 import 'package:flutter_banking_pay_responsive/models/enums.dart';
 import 'package:flutter_banking_pay_responsive/screens/activityInsights/activity_insights_screen.dart';
 import 'package:flutter_banking_pay_responsive/screens/cardScreen/card_screen.dart';
 import 'package:flutter_banking_pay_responsive/screens/homeScreen/home_screen.dart';
-import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:flutter/services.dart';
 import 'package:provider/provider.dart';
 
@@ -31,6 +28,12 @@ class SetupScreenState extends State<SetupScreen> {
     Consumer<CardScreen>(builder: (_, screen, __) => screen),
     Consumer<ActivityInsightsScreen>(builder: (_, screen, __) => screen),
   ];
+
+  late final MaterialYouNavigationBarCustom navBar =
+      MaterialYouNavigationBarCustom(
+    getCurrentIndex: getValueCurrentIndex,
+    callbackOnPress: callbackOnBottomNavigationPress,
+  );
 
   int _selectedIndex = 0;
   MenuState get selectedMenuState => MenuState.values[_selectedIndex];
@@ -72,91 +75,7 @@ class SetupScreenState extends State<SetupScreen> {
         body: Center(
           child: menuWidgets.elementAt(_selectedIndex),
         ),
-        bottomNavigationBar: buildMaterialYouNavigationBar(context),
-      ),
-    );
-  }
-
-  Widget buildMaterialYouNavigationBar(BuildContext context) {
-    return NavigationBarTheme(
-      data: NavigationBarThemeData(
-        height: kBottomNavigationHeight(context),
-        backgroundColor: Theme.of(context).colorScheme.background,
-        labelTextStyle:
-            //MaterialStateProperty.all(const TextStyle(fontSize: 14))
-            MaterialStateProperty.resolveWith<TextStyle>(
-          (Set<MaterialState> states) {
-            if (states.contains(MaterialState.selected)) {
-              return AppTextStyles.kNavigationBarSelectedText;
-            } else {
-              return AppTextStyles.kNavigationBarText;
-              // return null; // Use the component's default.
-            }
-          },
-        ),
-        labelBehavior: NavigationDestinationLabelBehavior.alwaysShow,
-        iconTheme: MaterialStateProperty.resolveWith<IconThemeData>(
-          (Set<MaterialState> states) {
-            if (states.contains(MaterialState.selected)) {
-              return const IconThemeData(size: kSmallIconSize * 1.3);
-            } else {
-              return const IconThemeData(size: kSmallIconSize);
-              // return null; // Use the component's default.
-            }
-          },
-        ),
-        indicatorColor: kSecondaryColor,
-      ),
-      child: NavigationBar(
-        selectedIndex: _selectedIndex,
-        onDestinationSelected: (int index) => changeSelectedMenu(index),
-        animationDuration: const Duration(seconds: 1, milliseconds: 700),
-        destinations: [
-          NavigationDestination(
-            icon: const Icon(FontAwesomeIcons.dollarSign),
-            label: S.of(context).navigationBar_first_title,
-          ),
-          NavigationDestination(
-            icon: const Icon(FontAwesomeIcons.solidCreditCard),
-            label: S.of(context).navigationBar_second_title,
-            selectedIcon: const Icon(FontAwesomeIcons.creditCard),
-          ),
-          NavigationDestination(
-            icon: const Icon(FontAwesomeIcons.chartLine),
-            label: S.of(context).navigationBar_third_title,
-          ),
-        ],
-      ),
-    );
-  }
-
-  Widget buildOldNavigationBar(BuildContext context) {
-    return SizedBox(
-      width: double.infinity,
-      height: kBottomNavigationHeight(context),
-      child: BottomNavigationBarTheme(
-        data: BottomNavigationBarThemeData(
-          type: BottomNavigationBarType.fixed,
-          backgroundColor: Theme.of(context).colorScheme.background,
-          selectedItemColor: Theme.of(context).colorScheme.primary,
-          selectedLabelStyle: AppTextStyles.kNavigationBarSelectedText,
-          enableFeedback: true,
-          elevation: 0,
-        ),
-        child: BottomNavigationBar(
-          currentIndex: _selectedIndex,
-          onTap: (int index) => changeSelectedMenu(index),
-          iconSize:
-              kMediumIconSize, // MenuState.pay == _selectedMenu ? 28 : 22,
-          items: const [
-            BottomNavigationBarItem(
-                icon: Icon(FontAwesomeIcons.dollarSign), label: 'Pay'),
-            BottomNavigationBarItem(
-                icon: Icon(FontAwesomeIcons.solidCreditCard), label: 'Cards'),
-            BottomNavigationBarItem(
-                icon: Icon(FontAwesomeIcons.chartLine), label: 'Insights'),
-          ],
-        ),
+        bottomNavigationBar: navBar,
       ),
     );
   }
@@ -180,7 +99,16 @@ class SetupScreenState extends State<SetupScreen> {
     });
   }
 
+  void callbackOnBottomNavigationPress(int index) {
+    // _selectedIndex
+    print('clicked! with $index');
+    changeSelectedMenu(index);
+  }
+
+  int getValueCurrentIndex() => _selectedIndex;
+
   void notifyPopDependencies() {
+    // TODO: create observer pattern
     // HomeScreen
     Provider.of<HomeScreen>(context, listen: false)
         .keyValueScreen
